Prepared data for Model
In this directory, you will find the processed input data that has been prepared specifically for use with the model. These files are ready to be used directly, allowing you to run the model without any additional preprocessing.

Simply load the provided data files into the codebase to reproduce the results or to test alternative models using the same inputs.